+++ shortbread/src/test/java/shortbread/ShortbreadTest.java
@@ -3,6 +3,7 @@ package shortbread;
 import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.Application;
+import android.content.Context;
 import android.content.Intent;
 import android.content.pm.ShortcutInfo;
 import android.content.pm.ShortcutManager;
@@ -11,6 +12,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
@@ -21,7 +23,6 @@ import java.util.List;
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
@@ -32,6 +33,8 @@ import static org.mockito.Mockito.when;
 public class ShortbreadTest {
 
     @Mock
+    private Context context;
+    @Mock
     private Activity activity;
     @Mock
     private Application application;
@@ -39,15 +42,13 @@ public class ShortbreadTest {
     private ShortcutManager shortcutManager;
     @Mock
     private Intent intent;
+    @Captor
+    ArgumentCaptor<Application.ActivityLifecycleCallbacks> lifecycleCallbacksCaptor;
 
     @Before
     public void setUp() {
-        Shortbread.shortcutsSet = false;
-        Shortbread.activityLifecycleCallbacksSet = false;
-
         MockitoAnnotations.openMocks(this);
-        when(activity.getApplicationContext()).thenReturn(application);
-        when(application.getApplicationContext()).thenReturn(application);
+        when(context.getApplicationContext()).thenReturn(application);
         when(application.getSystemService(ShortcutManager.class)).thenReturn(shortcutManager);
         when(activity.getIntent()).thenReturn(intent);
     }
@@ -55,89 +56,77 @@ public class ShortbreadTest {
     @Test
     @Config(sdk = 24)
     public void doesNothingBeforeApi25() {
-        Shortbread.create(activity);
-        verifyNoMoreInteractions(activity);
+        new Shortbread(context);
+        verifyNoMoreInteractions(context);
     }
 
     @Test
     public void usesApplicationContext() {
-        Shortbread.create(activity);
-        verify(activity).getApplicationContext();
+        new Shortbread(context);
+        verify(context).getApplicationContext();
         verify(application).getSystemService(ShortcutManager.class);
     }
 
     @Test
-    public void registersActivityLifecycleCallbacksOnce() {
-        Shortbread.create(activity);
+    public void registersActivityLifecycleCallbacks() {
+        new Shortbread(context);
         verify(application).registerActivityLifecycleCallbacks(any(Application.ActivityLifecycleCallbacks.class));
-
-        Shortbread.create(activity);
-        verify(application, times(1)).registerActivityLifecycleCallbacks(any(Application.ActivityLifecycleCallbacks.class));
     }
 
     @Test
-    public void setsEnabledShortcutsOnce() {
-        final List<ShortcutInfo> enabledShortcuts = ShortbreadGenerated.createShortcuts(activity).get(0);
+    public void setsEnabledShortcuts() {
+        final List<ShortcutInfo> enabledShortcuts = ShortbreadGenerated.createShortcuts(context).get(0);
 
-        Shortbread.create(activity);
+        new Shortbread(context);
         verify(shortcutManager).setDynamicShortcuts(enabledShortcuts);
-
-        Shortbread.create(activity);
-        verify(shortcutManager, times(1)).setDynamicShortcuts(enabledShortcuts);
     }
 
     @Test
-    public void setsDisabledShortcutsOnce() {
+    public void setsDisabledShortcuts() {
         final List<ShortcutInfo> disabledShortcuts = ShortbreadGenerated.createShortcuts(activity).get(1);
 
-        Shortbread.create(activity);
+        new Shortbread(context);
         @SuppressWarnings("rawtypes") ArgumentCaptor<List> captor = ArgumentCaptor.forClass(List.class);
         //noinspection unchecked
         verify(shortcutManager).disableShortcuts(captor.capture());
         assertEquals(disabledShortcuts.size(), captor.getValue().size());
-
-        Shortbread.create(activity);
-        //noinspection unchecked
-        verify(shortcutManager, times(1)).disableShortcuts(any(List.class));
-    }
-
-    @Test
-    public void intentHasExtra_callsMethodShortcut() {
-        when(intent.hasExtra("shortbread_method")).thenReturn(true);
-        ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut = null;
-
-        Shortbread.create(activity);
-        assertEquals(activity, ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
     }
 
     @Test
     public void intentDoesNotHaveExtra_doesNotCallMethodShortcut() {
         when(intent.hasExtra("shortbread_method")).thenReturn(false);
+
+        new Shortbread(context);
         ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut = null;
 
-        Shortbread.create(activity);
+        verify(application).registerActivityLifecycleCallbacks(lifecycleCallbacksCaptor.capture());
+        Application.ActivityLifecycleCallbacks lifecycleCallbacks = lifecycleCallbacksCaptor.getValue();
+
+        lifecycleCallbacks.onActivityCreated(activity, null);
+        assertNull(ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
+
+        lifecycleCallbacks.onActivityStarted(activity);
         assertNull(ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
     }
 
     @Test
     public void activityLifecycleListenerCallsMethodShortcut() {
         when(intent.hasExtra("shortbread_method")).thenReturn(true);
-        Shortbread.create(application);
+
+        new Shortbread(context);
         ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut = null;
 
-        ArgumentCaptor<Application.ActivityLifecycleCallbacks> captor = ArgumentCaptor.forClass(
-                Application.ActivityLifecycleCallbacks.class);
-        verify(application).registerActivityLifecycleCallbacks(captor.capture());
-        Application.ActivityLifecycleCallbacks activityLifecycleCallbacks = captor.getValue();
+        verify(application).registerActivityLifecycleCallbacks(lifecycleCallbacksCaptor.capture());
+        Application.ActivityLifecycleCallbacks lifecycleCallbacks = lifecycleCallbacksCaptor.getValue();
 
-        activityLifecycleCallbacks.onActivityCreated(activity, null);
+        lifecycleCallbacks.onActivityCreated(activity, null);
         assertNull(ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
 
-        activityLifecycleCallbacks.onActivityStarted(activity);
+        lifecycleCallbacks.onActivityStarted(activity);
         assertEquals(activity, ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
 
         ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut = null;
-        activityLifecycleCallbacks.onActivityStarted(activity);
+        lifecycleCallbacks.onActivityStarted(activity);
         assertNull(ShortbreadGenerated.activityThatWasPassedToCallMethodShortcut);
     }
 }